requirements.txt


cryptography~=37.0.4
django~=2.2.26
boto3~=1.26.38
django-environ~=0.9.0
django-storages~=1.12
psycopg2<2.9
waitress~=2.1.2
whitenoise~=6.0.0
djangorestframework~=3.13.1
django-bootstrap4~=22.1
django-allauth~=0.51.0
django-rest-auth~=0.9.5
django-extensions~=3.1.5
packaging
pyyaml~=6.0
pillow~=9.2.0
pytest==7.2.0
factory-boy==3.2.1
google-cloud-secret-manager==2.13.0
google-auth==2.12.0
google-cloud-storage==2.5.0
pipenv
django-import-export
drf-spectacular~=0.25.1
-e ./modules/privacy_policy
-e ./modules/terms_and_conditions
